Take response callback instead of action callback as argument
The notify() action creator now takes a more generic response callback
(instead of the button action callback) for other types of actions like
notification timeouts.

+++ src/thunks/remove-download.js
@@ -1,8 +1,6 @@
 import { removeDownload, notify, showDownload, hideDownload } from '../actions';
 import { getPartialDownloadPath, deleteFile } from '../utilities';
-import { NOTIFICATION_SHOW_DURATION } from '../constants';
 import pathExists from 'path-exists';
-import Timeout from 'await-timeout';
 
 export default function removeDownloadThunk(id) {
   return async (dispatch, getState) => {
@@ -10,25 +8,23 @@ export default function removeDownloadThunk(id) {
 
     const download = getState().downloads.byId[id];
 
-    const timeout = new Timeout();
-    timeout.set(NOTIFICATION_SHOW_DURATION).then(async () => {
-      if (download.status === 'canceled') {
-        dispatch(removeDownload(id));
-        const partialDownloadPath = getPartialDownloadPath(download);
-        if (await pathExists(partialDownloadPath)) {
-          await deleteFile(partialDownloadPath);
-        }
-      }
-    });
-
     dispatch(
       notify(
         'info',
         `Removed '${download.availableFilename}' from list`,
         'Undo',
-        () => {
-          dispatch(showDownload(id));
-          timeout.clear();
+        async (responseType) => {
+          if (responseType === 'undo') {
+            dispatch(showDownload(id));
+          } else if (responseType === 'timeout') {
+            if (download.status === 'canceled') {
+              dispatch(removeDownload(id));
+              const partialDownloadPath = getPartialDownloadPath(download);
+              if (await pathExists(partialDownloadPath)) {
+                await deleteFile(partialDownloadPath);
+              }
+            }
+          }
         }
       )
     );
